resume_generator/requirements.txt
asgiref==3.7.2
boto3==1.28.25
botocore==1.31.25
Django==4.2.4
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.2.0
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.4
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
